// crates/opsml_server/opsml_ui/src/lib/scripts/types.ts
export interface RunGraph {
  name: string;
  x_label: string;
  y_label: string;
  x: number[];
  y: number[] | Map<string, number[]>;
  graph_type: string;
  graph_style: string;
}
// ...
export interface ChartjsData {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  data: any;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  options: any;
  type: string;
}

export interface ChartjsLineDataset {
  label: string;
  data: number[];
  borderColor: string;
  backgroundColor: string;
  pointRadius?: number;
  fill?: boolean;
  tension?: number;
}

export interface ChartjsBarDataset {
  data: number[];
  borderColor: string[];
  backgroundColor: string[];
  borderWidth: number;
  borderRadius: number;
  borderSkipped: boolean;
}

export interface ScatterData {
  x: number;
  y: number;
}

export interface ChartjsScatterDataset {
  label: string;
  data: ScatterData[];
  borderColor: string;
  backgroundColor: string;
  pointRadius: number;
}